+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# cranko 0.0.2 (2020-08-22)
+
+- Add `github` subcommand with some useful utilities
+- Wire those up in the Cranko CI/CD pipeline
+- Add logging infrastructure and start making the CLI UI nicer
+- Work on adding a few more consistency checks
+- Fix changelog updating scheme
+
 # Version 0.0.1 (2020-08-19)
 
 Attempt to publish 0.0.1!